Put items collections to seperate users

+++ lib/repositories/items_repository.dart
@@ -1,10 +1,16 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-
+import 'package:firebase_auth/firebase_auth.dart';
 import '../models/item_model.dart';
 
 class ItemsRepository {
   Stream<List<ItemModel>> getItemsStream() {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
     return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('items')
         .orderBy('release_date')
         .snapshots()
@@ -23,12 +29,24 @@ class ItemsRepository {
   }
 
   Future<void> delete({required String id}) {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
     return FirebaseFirestore.instance.collection('items').doc(id).delete();
   }
 
   Future<ItemModel> get({required String id}) async {
-    final doc =
-        await FirebaseFirestore.instance.collection('items').doc(id).get();
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
+    final doc = await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
+        .collection('items')
+        .doc(id)
+        .get();
 
     return ItemModel(
       id: doc.id,
@@ -43,7 +61,15 @@ class ItemsRepository {
     String imageURL,
     DateTime releaseDate,
   ) async {
-    await FirebaseFirestore.instance.collection('items').add(
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not logged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
+        .collection('items')
+        .add(
       {
         'title': title,
         'image_url': imageURL,
